feat: `useScrollHandler` 생성해서 코드 분리
